# historian/backend/src/main/swagger/api-swagger.yml
swagger: '2.0'
info:
  description: REST API for logisland historian
  version: v1
  title: Logisland Data Historian
  contact:
    name: hurence
    email: support@hurence.com
host: localhost:8701
basePath: /api/v1
schemes:
  - http
  - https
paths:
...
  /tags:
    get:
      tags:
      - tag
      operationId: getAllTags
      parameters:
        - name: fq
          in: query
          description: filter query (lucene syntax like fq="labels:opc AND datasources:win32")
          required: false
          type: string
      summary:  get all saved tags
      description: retrieve all OPC tags
      produces:
        - application/json
      responses:
       '200':
         description: tags list
         schema:
           type: array
           items:
             $ref: '#/definitions/Tag'
       default:
         description: unexpected error
         schema:
           $ref: '#/definitions/Error'
  /tags/{itemId}:
     get:
       tags:
         - tag
       summary: get tag
       description: get the corresponding Tag
       operationId: getTag
       produces:
         - application/json
       parameters:
         - name: itemId
           in: path
           description: id of the tag to return
           required: true
           type: string
       responses:
         '200':
           description: tag
           schema:
             $ref: '#/definitions/Tag'
         '404':
            description: Tag resource not found
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
     delete:
       tags:
         - tag
       summary: delete tag
       description: remove the corresponding Tag
       operationId: deleteTag
       parameters:
         - name: itemId
           in: path
           description: id of the tag to be deleted
           required: true
           type: string
       responses:
         '404':
           description: Tag resource not found
         '400':
           description: Invalid ID supplied
         '200':
           description: Tag successfully removed
           schema:
             $ref: '#/definitions/Tag'
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
     put:
       tags:
        - tag
       summary:  update tag
       description: update an existing tag
       operationId: updateTag
       parameters:
         - name: itemId
           in: path
           description: itemId to be updated
           required: true
           type: string
         - name: tag
           in: body
           description: new Tag definition
           required: true
           schema:
             $ref: '#/definitions/Tag'
       responses:
        '404':
          description: Tag resource not found
        '400':
          description: Invalid ID supplied
        '200':
         description: Tag successfuly updated
         schema:
           $ref: '#/definitions/Tag'
     post:
       tags:
        - tag
       summary:  create new tag
       description:  store a new tag
       operationId: addTagWithId
       parameters:
         - name: body
           in: body
           description: Tag resource to add
           required: true
           schema:
             $ref: '#/definitions/Tag'
         - name: itemId
           in: path
           description: itemId to
           required: true
           type: string
       responses:
         '400':
           description: Invalid ID supplied
         '200':
           description: Tag successfuly created
           schema:
             $ref: '#/definitions/Tag'
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
  /tags/{itemId}/measures:
     get:
       tags:
         - tag
         - measure
       summary: get tag measures
       description: get the corresponding Tag measures between start and end time
       operationId: getTagMeasures
       produces:
         - application/json
       parameters:
         - name: itemId
           in: path
           description: id of the tag
           required: true
           type: string
         - name: start
           in: query
           description: |
            date de début (borne inf) peut-être exprimée sous les formats suivants :
              timestamp : 4578965
              date-time : 2015-11-25T12:06:57.330Z
              relatif   : NOW-30DAYS
           required: false
           type: string
         - name: end
           in: query
           description: |
             date de fin (borne sup) peut-être exprimée sous les formats suivants :
               timestamp : 4578965
               date-time : 2015-11-25T12:06:57.330Z
               relatif   : NOW-30DAYS
           required: false
           type: string
         - name: functions
           in: query
           description: |
              Multiple analyses, aggregations, and transformations are allowed per query.
              If so, Chronix will first execute the transformations in the order they occur.
              Then it executes the analyses and aggregations on the result of the chained transformations.
              For example the query:

                max;min;trend;movavg:10,minutes;scale:4

              is executed as follows:

                Calculate the moving average
                Scale the result of the moving average by 4
                Calculate the max, min, and the trend based on the prior result.
           required: false
           type: string
         - name: no_values
           in: query
           description: will retrieve only function values, no data points
           required: false
           type: boolean
           default: false
       responses:
         '200':
           description: tag
           schema:
             $ref: '#/definitions/Measures'
         '404':
            description: Tag resource not found
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
  /tags/{itemId}/stats:
     get:
       tags:
         - tag
         - measure
       summary: get tag measures stats
       description: get the corresponding Tag mesures for last chunk
       operationId: getTagStats
       produces:
         - application/json
       parameters:
         - name: itemId
           in: path
           description: id of the tag
           required: true
           type: string
       responses:
         '200':
           description: tag
           schema:
             $ref: '#/definitions/Measures'
         '404':
            description: Tag resource not found
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
  /tags/measures:
     post:
       tags:
         - tag
         - measure
       summary: post tag measures
       description: post some new values
       operationId: postTagMeasures
       produces:
         - application/json
       parameters:
         - name: content
           in: formData
           description: the csv file content
           required: true
           type: file
         - name: csv_delimiter
           in: query
           description: the csv file content
           required: false
           type: string
           default: ';'
         - name: date_format
           in: query
           description: |
            valid values LONG (ms since 1970),
              INSTANT (default java 8 instant),
              'SDF-FORMAT' e.g dd.MM.yyyy HH:mm:ss.SSS
           required: false
           type: string
           default: 'dd.MM.yyyy HH:mm:ss.SSS'
         - name: number_format
           in: query
           description: valid values ENGLISH, GERMAN
           required: false
           type: string
           default: 'ENGLISH'
         - name: attribute_fields
           in: query
           required: false
           type: string
         - name: clean_import
           in: query
           description: will discard all previously loaded data (use it with great care)
           required: false
           type: boolean
           default: false
         - name: points_by_chunk
           in: query
           description: the number of points by chunk
           required: false
           type: integer
           format: int32
       responses:
         '200':
           description: BulkLoad result
           schema:
             $ref: '#/definitions/BulkLoad'
         '404':
            description: Tag resource not found
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
  /tags/measures/generator:
     post:
       tags:
         - tag
         - measure
       summary: post tag measures simulation
       description: post some new values according to a simulation
       operationId: postTagMeasuresGenerator
       produces:
         - application/json
       parameters:
         - name: config
           in: formData
           description: the json config file
           required: true
           type: file
         - name: attribute_fields
           in: query
           required: false
           type: string
         - name: clean_import
           in: query
           description: will discard all previously loaded data (use it with great care)
           required: false
           type: boolean
           default: false
       responses:
         '200':
           description: BulkLoad result
           schema:
             $ref: '#/definitions/BulkLoad'
         '404':
            description: Tag resource not found
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'

  # Selection API
...
definitions:
  BulkLoad:
    type: object
    properties:
      start_time:
        type: string
      generation_duration:
        type: integer
        format: int32
        default: 0
      import_duration:
        type: integer
        format: int32
      num_metrics_imported:
        type: integer
        format: int32
      num_points_imported:
        type: integer
        format: int64
      metrics:
         type: array
         items:
          type: string
...
  Tag:
    type: object
    description: a Tag is an identifier to an OPC value
    required:
      - item_name
      - datasource_id
      - data_type
    properties:
      record_type:
        type: string
        default: 'tag'
      id:
        type: string
        default: 'mySweetUniqueId'
      datasource_id:
        type: string
      domain:
        type: string
      server:
        type: string
      group:
        type: string
      tag_name:
        type: string
      labels:
        description: can be used to group tags around a theme
        type: array
        items:
          type: string
      data_type:
        type: string
        default: float
        enum:
        - int
        - long
        - float
        - double
        - string
        - array
        - null
        - bytes
        - boolean
      description:
        type: string
      text:
        description: ctach all field
        type: array
        items:
          type: string
      creation_date:
        description: timestamp
        type: integer
        format: int64
      last_modification_date:
        description: timestamp
        type: integer
        format: int64
      last_polling_date:
        description: timestamp
        type: integer
        format: int64
      update_rate:
        description: polling delay in ms
        default: 1000
        type: integer
      min_numeric_value:
        type: number
        format: dumble
      max_numeric_value:
        type: number
        format: double
      last_numeric_value:
        type: number
        format: double
      last_quality:
        type: integer
  Measures:
    type: object
    description: holds a double array of values for a Tag between start & stop
    properties:
      name:
        type: string
      start:
        type: integer
        format: int64
      end:
        type: integer
        format: int64
      query_duration:
        type: integer
        format: int64
      quality:
        type: number
        format : double
      num_chunks:
        type: integer
        format: int64
      num_points:
        type: integer
        format: int32
      timestamps:
        type: array
        items:
          type: integer
          format: int64
      values:
        type: array
        items:
          type: number
          format: double
      functions:
        type: array
        items:
          $ref : "#/definitions/Function"
  Function:
    type: object
    properties:
      name:
        type: string
      value:
        type: number
        format: double
...
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string